pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]==0.109.0",
    "firebase-admin==6.4.0",
    "psycopg[binary]==3.1.18",
    "pydantic==2.6.0",
    "pydantic-settings==2.1.0",
    "python-dotenv==1.0.1",
    "sqlmodel==0.0.14",
    "uvicorn==0.27.0",
]
